(chore) removed the spacing in between buttons and tabs
