+++ packages/esm-shr-app/src/referrals-wrap.tsx
@@ -1,7 +1,6 @@
 import React, { useState } from 'react';
 import { ReferralsHeader } from './header/referrals-header.component';
 import ReferralTabs from './referrals/referral-tabs/referrals-tabs.component';
-import MetricsHeader from './header/referrals-button-header.component';
 import { mutate } from 'swr';
 import { pullFacilityReferrals } from './referrals/refferals.resource';
 
@@ -25,8 +24,7 @@ const ReferralWrap: React.FC = () => {
   return (
     <div className={`omrs-main-content`}>
       <ReferralsHeader />
-      <MetricsHeader pullReferrals={pullReferrals} isLoadingFacilityReferrals={isLoadingFacilityReferrals} />
-      <ReferralTabs isLoadingFacilityReferrals={isLoadingFacilityReferrals} />
+      <ReferralTabs isLoadingFacilityReferrals />
     </div>
   );
 };
